requirements.txt
PyQt6>=6.4.0
xxhash>=3.0.0
Pillow>=9.0.0
imagehash>=4.3.1
send2trash>=1.8.0
pathlib